test(settings): split the settings spec along the facade seams
settings.component.spec.ts was 1516 lines and the last settings file left in
the max-lines baseline. The behaviour that moved into facades now has its own
specs, driven directly instead of through the rendered page.

- settings-app-update.facade.spec.ts: status polling/retry, bridge actions,
  release notes dialog, version messaging (+ dispose, up-to-date message)
- settings-epg.facade.spec.ts: refresh, clear flow (+ blank-source skip,
  refresh-all, post-save re-fetch)
- settings-playlist-reset.facade.spec.ts: summary, dialog, Electron progress,
  browser fallback (+ failure snackbar)
- settings-backup.facade.spec.ts: desktop export, browser download fallback
  (+ failure snackbar)
- settings.component.spec.ts keeps the page shell and runtime capabilities;
  settings.component.form.spec.ts takes hydration, section outputs, dashboard
  controls and submit
- shared TestBed fixtures live in settings-test-harness.stub.ts (.stub.ts so
  the app build skips it, matching electron.service.stub.ts)

104 settings tests (was 94); every file is now under the 400-line limit, so
settings.component.spec.ts leaves the baseline.
