+++ tools/eslint/max-lines-baseline.mjs
@@ -42,7 +42,6 @@ export const maxLinesBaseline = [
     'apps/web-e2e/src/xtream.e2e.ts',
     'apps/web/src/app/services/electron.service.ts',
     'apps/web/src/app/services/pwa.service.ts',
-    'apps/web/src/app/settings/settings.component.spec.ts',
     'apps/xtream-mock-server/src/app/generators/marketing.generator.ts',
     'libs/epg/data-access/src/lib/epg.service.spec.ts',
     'libs/epg/data-access/src/lib/epg.service.ts',
